# PERF: do NPY_NAT check inside get_datetime64_nanos